# src/main/resources/templates/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Connect new sites to the OrganiCity platform, as well as perform the management of its related attributes",
		"version": "1.0.0",
		"title": "OrganiCity Site Management API",
		"contact": {"email": "etienne.gandrille@cea.fr"}
	},
	"host": "{{host}}:{{exposedport}}",
	"basePath": "/v1",
...
	"schemes": ["{{protocol}}"],
	"paths": {
...
		"/sites/{siteName}/services": {
			"get": {
				"tags": ["site"],
				"summary": "Finds the services of a site",
				"operationId": "getServices",
				"produces": ["application/json"],
				"parameters": [{
					"name": "siteName",
					"in": "path",
					"description": "Site name",
					"required": true,
					"type": "string"
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:site-{siteName}-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "An array of OC services",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/servicedesc"
							}
						}
					},
					"400": {
						"description": "Invalid name supplied"
					},
					"401": {
						"description": "Authorization error"
					},
					"404": {
						"description": "Site not found"
					}
				}
			},
			"post": {
				"tags": ["site"],
				"summary": "Add a new service in a given site",
				"operationId": "addService",
				"consumes": [
					"application/json"
				],
				"produces": ["application/json"],
				"parameters": [{
					"name": "siteName",
					"in": "path",
					"description": "Site name",
					"required": true,
					"type": "string"
				}, {
					"in": "body",
					"name": "site",
					"description": "Service object that needs to be added to the facility",
					"required": true,
					"schema": {
						"$ref": "#/definitions/service"
					}
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:site-{siteName}-admin",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Service added",
						"schema": {
							"$ref": "#/definitions/servicedesc"
						}
					},
					"400": {
						"description": "Malformed or invalid json content, or entity with an already existing name supplied"
					},
					"401": {
						"description": "Authorization error"
					},
					"404": {
						"description": "Site not found"
					},
					"405": {
						"description": "Invalid json content (i.e. missing fields) or already existing service"
					}
				}
			}
		},
		"/sites/{siteName}/services/{serviceName}": {
			"delete": {
				"tags": ["site"],
				"summary": "Deletes a service",
				"operationId": "deleteService",
				"produces": ["application/json"],
				"parameters": [{
					"name": "siteName",
					"in": "path",
					"description": "Site name",
					"required": true,
					"type": "string"
				}, {
					"name": "serviceName",
					"in": "path",
					"description": "Service name to be deleted",
					"required": true,
					"type": "string"
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:site-{siteName}-admin",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Service removed"
					},
					"400": {
						"description": "Invalid siteName or serviceName supplied"
					},
					"401": {
						"description": "Authorization error"
					},
					"404": {
						"description": "Site or service in site not found"
					}
				}
			},
			"get": {
				"tags": ["site"],
				"summary": "Find service by name",
				"operationId": "getServiceByName",
				"produces": ["application/json"],
				"parameters": [{
					"name": "siteName",
					"in": "path",
					"description": "Site name",
					"required": true,
					"type": "string"
				}, {
					"name": "serviceName",
					"in": "path",
					"description": "Name of service that needs to be fetched",
					"required": true,
					"type": "string"
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:site-{siteName}-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/servicedesc"
						}
					},
					"400": {
						"description": "Invalid siteName or serviceName supplied"
					},
					"401": {
						"description": "Authorization error"
					},
					"404": {
						"description": "Site or service in site not found"
					}
				}
			},
			"put": {
				"tags": ["site"],
				"summary": "Updates a service",
				"operationId": "updateService",
				"consumes": [
					"application/json"
				],
				"produces": ["application/json"],
				"parameters": [{
					"name": "siteName",
					"in": "path",
					"description": "Site name",
					"required": true,
					"type": "string"
				}, {
					"name": "serviceName",
					"in": "path",
					"description": "Name of service that needs to be updated. It MUST be the same name as the one in the body",
					"required": true,
					"type": "string"
				}, {
					"in": "body",
					"name": "site",
					"description": "Service that needs to be updated",
					"required": true,
					"schema": {
						"$ref": "#/definitions/serviceupdate"
					}
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:site-{siteName}-admin",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Service updated",
						"schema": {
							"$ref": "#/definitions/servicedesc"
						}
					},
					"400": {
						"description": "Malformed or invalid json content (i.e. missing fields), or invalid name supplied"
					},
					"401": {
						"description": "Authorization error"
					},
					"404": {
						"description": "Site or service not found"
					}
				}
			}
		},
...
		"/dictionary/assettypes": {
			"get": {
				"tags": ["dictionary"],
				"summary": "Finds all the asset types",
				"operationId": "getAssetTypes",
				"produces": ["application/json"],
				"parameters": [{
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:dictionary-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "An array of OC asset types",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/assettypedesc"
							}
						}
					},
					"401": {
						"description": "Authorization error"
					}
				}
			}
		},
		"/dictionary/assettypes/{name}": {
			"get": {
				"tags": ["dictionary"],
				"summary": "Find asset type by name",
				"operationId": "getAssetTypeByName",
				"produces": ["application/json"],
				"parameters": [{
					"name": "name",
					"in": "path",
					"description": "asset type name",
					"required": true,
					"type": "string"
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:dictionary-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/assettypedesc"
						}
					},
					"401": {
						"description": "Authorization error"
					},
					"404": {
						"description": "Asset Type not found"
					}
				}
			}
		},
		"/dictionary/attributetypes": {
			"get": {
				"tags": ["dictionary"],
				"summary": "Finds all the attribute types",
				"operationId": "getAttributeTypes",
				"produces": ["application/json"],
				"parameters": [{
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:dictionary-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "An array of OC attribute types",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/attributetypedesc"
							}
						}
					},
					"401": {
						"description": "Authorization error"
					}
				}
			}
		},
		"/dictionary/attributetypes/{name}": {
			"get": {
				"tags": ["dictionary"],
				"summary": "Find attribute type by name",
				"operationId": "getAttributeTypeByName",
				"produces": ["application/json"],
				"parameters": [{
					"name": "name",
					"in": "path",
					"description": "attribute type name",
					"required": true,
					"type": "string"
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:dictionary-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/attributetypedesc"
						}
					},
					"401": {
						"description": "Authorization error"
					},
					"404": {
						"description": "Attribute Type not found"
					}
				}
			}
		},
...
		"/dictionary/datatypes": {
			"get": {
				"tags": ["dictionary"],
				"summary": "Finds all the data types",
				"operationId": "getDataTypes",
				"produces": ["application/json"],
				"parameters": [{
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:dictionary-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "An array of OC data types",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/datatypedesc"
							}
						}
					},
					"401": {
						"description": "Authorization error"
					}
				}
			}
		},
		"/dictionary/datatypes/{name}": {
			"get": {
				"tags": ["dictionary"],
				"summary": "Find data type by name",
				"operationId": "getDataTypeByName",
				"produces": ["application/json"],
				"parameters": [{
					"name": "name",
					"in": "path",
					"description": "data type name",
					"required": true,
					"type": "string"
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:dictionary-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/datatypedesc"
						}
					},
					"401": {
						"description": "Authorization error"
					},
					"404": {
						"description": "Data Type not found"
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"service": {
			"type": "object",
			"required": [
				"name"
			],
			"properties": {
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC service"
				},
				"description": {
					"type": "string",
					"description": "optional description"
				},
				"related": {
					"type": "string",
					"description": "extension point for storing structured or unstructured information"
				}
			}
		},
		"serviceupdate": {
			"type": "object",
			"properties": {
				"description": {
					"type": "string",
					"description": "optional description"
				},
				"related": {
					"type": "string",
					"description": "extension point for storing structured or unstructured information"
				}
			}
		},
		"servicedesc": {
			"type": "object",
			"required": [
				"urn",
				"name",
				"created",
				"updated"
			],
			"properties": {
				"urn": {
					"type": "string",
					"description": "unique urn representing a specific OC service"
				},
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC service"
				},
				"description": {
					"type": "string",
					"description": "Optional description"
				},
				"related": {
					"type": "string",
					"description": "extension point for storing structured or unstructured information"
				},
				"created": {
					"type": "integer",
					"format": "int64",
					"description": "service registration date"
				},
				"updated": {
					"type": "integer",
					"format": "int64",
					"description": "service last update date"
				}
			}
		},
...
		"assettypedesc": {
			"type": "object",
			"required": [
				"urn",
				"name",
				"description",
				"related",
				"attributes"
			],
			"properties": {
				"urn": {
					"type": "string",
					"description": "unique urn representing a specific OC asset type"
				},
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC asset type"
				},
				"description": {
					"type": "string",
					"description": "Optional description"
				},
				"related": {
					"type": "string",
					"description": "extension point for storing structured or unstructured information"
				},
				"attributes": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/attributetypedesc"
					}
				}
			}
		},
		"attributetypedesc": {
			"type": "object",
			"required": [
				"urn",
				"name",
				"description",
				"related",
				"units"
			],
			"properties": {
				"urn": {
					"type": "string",
					"description": "unique urn representing a specific OC asset type"
				},
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC asset type"
				},
				"description": {
					"type": "string",
					"description": "Optional description"
				},
				"related": {
					"type": "string",
					"description": "extension point for storing structured or unstructured information"
				},
				"units": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/unitdesc"
					}
				}
			}
		},
		"unitdesc": {
			"type": "object",
			"required": [
				"urn",
				"name",
				"description",
				"related",
				"datatype"
			],
			"properties": {
				"urn": {
					"type": "string",
					"description": "unique urn representing a specific OC unit"
				},
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC unit"
				},
				"description": {
					"type": "string",
					"description": "Optional description"
				},
				"related": {
					"type": "string",
					"description": "extension point for storing structured or unstructured information"
				},
				"datatype": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/datatypedesc"
					}
				}
			}
		},
		"datatypedesc": {
			"type": "object",
			"required": [
				"urn",
				"name"
			],
			"properties": {
				"urn": {
					"type": "string",
					"description": "unique urn representing a specific OC data type"
				},
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC data type"
				}
			}
		},
...
	}
}